+++ lihil/endpoint/returns.py
@@ -16,7 +16,7 @@ from lihil.constant.status import code as get_status_code
 from lihil.errors import InvalidStatusError, NotSupportedError, StatusConflictError
 from lihil.interface import MISSING, CustomEncoder, IEncoder, Maybe, Record, is_provided
 from lihil.interface.marks import RESP_RETURN_MARK, ResponseMark, extract_resp_type
-from lihil.plugins.auth.jwt import jwt_encoder_factory
+from lihil.auth.jwt import jwt_encoder_factory
 from lihil.utils.json import encode_json, encode_text
 from lihil.utils.typing import get_origin_pro, is_union_type